Move ImageMetadata into its own file
Cleans up webui, enables webui_streamlit et al to use it as well.
